requirements.txt
fastapi>=0.111.0,<0.112.0
uvicorn[standard]>=0.30.0,<0.31.0
supabase>=2.0.0,<3.0.0
pydantic>=2.6.0,<3.0.0
pydantic-settings>=2.2.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
google-generativeai>=0.6.0,<0.7.0
google-genai>=0.3.0,<0.4.0
# Dependência para transcrição de áudio (requer FFmpeg instalado):
faster-whisper>=1.0.0
yt-dlp>=2024.8.6
youtube-transcript-api>=0.6.2,<0.7.0
httpx>=0.27.0,<0.28.0
psutil>=5.9.0,<6.0.0
# R2/S3 storage client
boto3>=1.34.0,<2.0.0
